+++ Orion/orion-server/src/main/java/com/sesnu/orion/web/model/Order.java
@@ -43,7 +43,7 @@ public class Order {
 	private Integer qtyPerPack;
 	
 	@Column(name = "pck_per_cont")
-	private Integer pckPerCont;
+	private Double pckPerCont;
 	
 	@Column(name = "cont_size")
 	private Integer contSize;
@@ -193,11 +193,11 @@ public class Order {
 		this.qtyPerPack = qtyPerPack;
 	}
 
-	public Integer getPckPerCont() {
+	public Double getPckPerCont() {
 		return pckPerCont;
 	}
 
-	public void setPckPerCont(Integer pckPerCont) {
+	public void setPckPerCont(Double pckPerCont) {
 		this.pckPerCont = pckPerCont;
 	}
