+++ Orion/orion-server/src/main/java/com/sesnu/orion/web/model/SalesView.java
@@ -37,7 +37,7 @@ public class SalesView {
 	private int qtyPerPack;
 	
 	@Column(name = "pck_per_cont")
-	private int pckPerCont;
+	private Double pckPerCont;
 	
 	@Column(name = "cont_size")
 	private int contSize;
@@ -125,11 +125,11 @@ public class SalesView {
 		this.qtyPerPack = qtyPerPack;
 	}
 
-	public int getPckPerCont() {
+	public Double getPckPerCont() {
 		return pckPerCont;
 	}
 
-	public void setPckPerCont(int pckPerCont) {
+	public void setPckPerCont(Double pckPerCont) {
 		this.pckPerCont = pckPerCont;
 	}
